# docs/swagger.yaml
basePath: /
definitions:
  db.TableResponse:
    properties:
      data:
        items:
          additionalProperties: true
          type: object
        type: array
      meta:
        $ref: '#/definitions/db.tableResponseMeta'
    type: object
  db.property:
    properties:
      desc:
        type: boolean
      name:
        type: string
      remap:
        type: string
      table:
        type: string
    type: object
  db.tableResponseMeta:
    properties:
      available_props:
        items:
          $ref: '#/definitions/db.property'
        type: array
      count:
        type: integer
      included_props:
        items:
          $ref: '#/definitions/db.property'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  tables.NodeTag:
    properties:
      created:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      node_id:
        type: string
      tag_attach_data:
        items:
          type: integer
        type: array
      tag_id:
        type: string
      updated_at:
        type: string
    type: object
...
  tables.User:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      email_log_level:
        type: string
      email_notifications:
        type: string
      first_name:
        type: string
      id:
        type: integer
      im_log_level:
        type: string
      im_notifications:
        type: string
      im_notifications_delay:
        type: integer
      im_type:
        type: string
      im_username:
        type: string
      last_name:
        type: string
      lock_filter:
        type: string
      password:
        type: string
      phone_work:
        type: string
      quota_app:
        type: integer
      quota_docker_registries:
        type: integer
      quota_org_group:
        type: integer
      registration_id:
        type: string
      registration_key:
        type: string
      reset_password_key:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: collector-api-contact@opensvc.com
    name: OpenSVC SAS
    url: https://www.opensvc.com
  description: Organization clusters, nodes, services and more.
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: OpenSVC collector API
  version: "1.0"
paths:
...
  /nodes/{id}/candidate_tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List existing tags not already attached to a node
      tags:
      - tags
  /nodes/{id}/tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List tags attached to a node
      tags:
      - tags
  /nodes/{id}/tags/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.NodeTag'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Show a tag attachment to a node
      tags:
      - tags
      - nodes
  /nodes/tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List tags attachments to nodes
      tags:
      - tags
  /nodes/tags/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.NodeTag'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Show a tag attachment to a node
      tags:
      - tags
      - nodes
...
  /tags/{id}/nodes:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List nodes having a specific tag
      tags:
      - tags
...
  /users:
    get:
      consumes:
      - application/json
      description: List users
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: |-
        The user must be in the UserManager privilege group to modify tiers users properties.
        The user must be in the SelfManager privilege group to modify its user properties.
      parameters:
      - description: list of users to create or update
        in: body
        name: users
        required: true
        schema:
          items:
            $ref: '#/definitions/tables.User'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create or update users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by index, email or login
      parameters:
      - description: the index of the entry in database, or email, or login name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.User'
            type: array
        "204":
          description: No Content
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Show a user by index, email or login
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.User'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Show a user
      tags:
      - users
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"